requirements.txt
speedtest-cli
ifcfg